refactor(ReservationCompletedActivity): 권한 유도 코드 이동
