+++ app/src/main/java/woowacourse/movie/view/moviemain/MovieMainActivity.kt
@@ -1,16 +1,7 @@
 package woowacourse.movie.view.moviemain
 
-import android.Manifest
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.content.Context
-import android.content.SharedPreferences
-import android.content.pm.PackageManager
-import android.os.Build
 import android.os.Bundle
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.commit
 import com.google.android.material.bottomnavigation.BottomNavigationView
@@ -18,16 +9,11 @@ import woowacourse.movie.R
 import woowacourse.movie.view.moviemain.movielist.MovieListFragment
 import woowacourse.movie.view.moviemain.reservationlist.ReservationListFragment
 import woowacourse.movie.view.moviemain.setting.SettingFragment
-import woowacourse.movie.view.seatselection.AlarmReceiver
 
 class MovieMainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_movie_main)
-
-        createChannel()
-
-        requestNotificationPermission()
 
         val navigation = findViewById<BottomNavigationView>(R.id.navigation_view)
 
@@ -55,45 +41,6 @@ class MovieMainActivity : AppCompatActivity() {
         supportFragmentManager.commit {
             setReorderingAllowed(true)
             replace(R.id.fragment_container_view, fragment)
-        }
-    }
-
-    private fun createChannel() {
-        val name = "Reservation Notification"
-        val channel = NotificationChannel(
-            AlarmReceiver.CHANNEL_ID,
-            name,
-            NotificationManager.IMPORTANCE_DEFAULT,
-        )
-        val notificationManager =
-            getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        notificationManager.createNotificationChannel(channel)
-    }
-
-    private fun requestNotificationPermission() {
-        if (ContextCompat.checkSelfPermission(
-                this,
-                Manifest.permission.POST_NOTIFICATIONS,
-            ) != PackageManager.PERMISSION_GRANTED
-        ) {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-                if (!shouldShowRequestPermissionRationale(Manifest.permission.POST_NOTIFICATIONS)) {
-                    requestPermissionLauncher.launch(Manifest.permission.POST_NOTIFICATIONS)
-                }
-            }
-        }
-    }
-
-    private val requestPermissionLauncher = registerForActivityResult(
-        ActivityResultContracts.RequestPermission(),
-    ) { isGranted: Boolean ->
-        val sharedPreferences =
-            this.getSharedPreferences(SettingFragment.ALARM_SETTING, Context.MODE_PRIVATE)
-        val editor: SharedPreferences.Editor = sharedPreferences.edit()
-        if (isGranted) {
-            editor.putBoolean(SettingFragment.IS_ALARM_ON, true).apply()
-        } else {
-            editor.putBoolean(SettingFragment.IS_ALARM_ON, false).apply()
         }
     }
 }
